API Reference

API Reference

API Reference

Defining Command Line Option Spec.

CliOptionSpec

Option

FlagOption

CounterOption

Positional

HelpOption

OptionGroup

MutexGroup

Parsing Arguments

parse_args

ParseResult

update_defaults

Error Handling

print_usage

CliOptionError

Internal Types

AbstractOption

AbstractOptionGroup

Defining Command Line Option Spec.

CliOptionSpec

CliOptions.CliOptionSpec — Type.

CliOptionSpec(options::AbstractOption...;
              program = PROGRAM_FILE,
              use_double_dash = true,
              onhelp = 0,
              onerror = 1)

A type representing a command line option specification.

program parameter is used for the program name which appears in help (usage) message. If omitted, Base.PROGRAM_FILE will be used.

If use_double_dash parameters is true, no argument after double dash (--) will be recognized as an option. In this case, the double dash itself will not parsed as an option nor a positional argument. Note that only the first double dash is treated specially so double dashes which appeares after it will be recognized as positional arguments. This is especially useful for programs which launches another program using command line arguments given to itself.

onhelp parameter controls what to do if a HelpOption was used. It can be either:

An Integer

The running program will print help message and exit using it as the status code.

nothing

Nothing happens. In this case, the HelpOption is treated just like a FlagOption so you can examine whether it was used or not by examining ParseResult using its name.

A function which takes no arguments

Do whatever you want in the function.

The default value is 0.

onerror parameter controls the action when an error was detected on parsing arguments. Available choices are:

An Integer

The running program will print an error message along with a help message and exit with the status code.

nothing

Ignore errors. Note that error messages are stored in _errors field of the returning ParseResult so you can examine them later.

A function which takes an error message

Example 1) onerror = (msg) -> (@warn msg) ... Warn the error but continue processing

Example 2) onerror = error ... Throw ErrorException using Base.exit, instead of exiting

The default value is 1.

Example: Using a function for onhelp parameter

using CliOptions

spec = CliOptionSpec(
    HelpOption(),
    onhelp = () -> begin
        print_usage(spec, verbose = false)
        # exit(42)  # Use exit() to let the program exit inside parse_args()
    end,
    program = "onhelptest.jl"
)
options = parse_args(spec, ["-h"])  # The program does not exit here
println(options.help)

# output

Usage: onhelptest.jl [-h]
true

Example: Using a function for onerror parameter

using CliOptions

spec = CliOptionSpec(
    Option("--required-argument"),
    onerror = (msg) -> println("Warning: $msg"),
)
options = parse_args(spec, String[])
println(repr(options.required_argument))

# output

Warning: Option "--required-argument" must be specified
missing

source

Option

CliOptions.Option — Type.

Option([type=String,] primary_name::String, secondary_name::String = "";
       default = missing, until = nothing, requirement = nothing, help = "")

Type representing a command line option whose value is a following argument. Two forms of option notations are supported:

Short form (e.g.: -n 42)

Starting with a dash, one character follows it

A following command line argument will be the option's value

Long form (e.g.: --foo-bar)

Starting with two dash, dash-separated words follow them

Value can be specified as one of the two forms below:

--foo-bar value; a following command line argument becomes the option's value

--foo-bar=value; characters after an equal sign following the option name becomes the option's value

An Option can have two names. primary_name is typically a short form notation and is also used to express the option in a usage message or error messages. secondary_name is typically a long form notation and is also used to generate a value name in a usage message. For example, if names of an option are -n and --foo-bar, it will appear in a usage message as -n FOO_BAR. If you want to define an option which have only a long form notation, specify it as primary_name and omit secondary_name.

If type parameter is set, option values will be converted to the type inside parse_args and will be stored in returned ParseResult.

default parameter is used when parse_args does not see the option in the given command line arguments. If a value other than missing was specified, it will be the option's value. If it's missing, absense of the option is considered as an error; in other word, the option becomes a required option. The default value of default parameter is missing.

If until parameter is specified, following arguments will be collected into a vector to be the option's value until an argument which is or one of the until parameter appears. In this case, type of the option's value will be Vector{T} where T is the type specified with type parameter. until parameter can be a string, a vector or tuple of strings, or nothing. Default value is nothing; no collection will be done.

requirement determines how to validate the option's value. If the option's value does not meet the requirement, it's considered an error. requirement can be one of:

nothing

Any value will be accepted

A list of acceptable values

Arguments which matches one of the values will be accepted

Any iterable can be used to specify acceptable values

Arguments will be converted to the specified type and then compared to each element of the list using function ==

A Regex

Arguments which matches the regular expression will be accepted

Pattern matching will be done for unprocessed input string, not type converted one

A custom validator function

It validates command line arguments one by one

It can return a Bool which indicates whether a given argument is acceptable or not

It also can return a String describing why a given command line argument is NOT acceptable, or an empty String if it is acceptable

If you want an option which does not take a command line argument as its value, see FlagOption and CounterOption

source

FlagOption

CliOptions.FlagOption — Type.

FlagOption(primary_name::String, secondary_name::String = "";
           negators::Union{String,Vector{String}} = String[],
           help = "",
           negator_help = "")

FlagOption represents a so-called "flag" command line option. An option of this type takes no value and whether it was specified becomes a boolean value.

source

CounterOption

CliOptions.CounterOption — Type.

CounterOption([type=Int,] primary_name::String, secondary_name::String = "";
              decrementers::Union{String,Vector{String}} = String[],
              default::Signed = 0,
              help::String = "",
              decrementer_help = "")

A type represents a flag-like command line option. Total number of times a CounterOption was specified becomes the option's value.

source

Positional

CliOptions.Positional — Type.

Positional([type=String,] singular_name, plural_name = "";
           multiple = false, requirement = nothing,
           default = missing, help = "")

Positional represents a command line argument which are not an option name nor an option value.

requirement determines how to validate positional arguments. See explanation of Option for more detail.

source

HelpOption

CliOptions.HelpOption — Type.

HelpOption(names = ("-h", "--help"); [help::String])

Options for printing help (usage) message.

The default value of names are -h and --help. If you do not like to have -h for printing help message, just give --help for names parameter (i.e.: HelpOption("--help"; ...)).

The default behavior for a help option is printing help message and exiting. If you do not like this behavior, use onhelp parameter on constructing CliOptionSpec.

source

OptionGroup

CliOptions.OptionGroup — Type.

OptionGroup(options::AbstractOption...; name::String = "")

OptionGroup contains one or more AbstractOptions and accepts command line arguments if one of the options is accepted. In other word, this is an OR operator for AbstractOptions.

source

MutexGroup

CliOptions.MutexGroup — Type.

MutexGroup(options::AbstractOption...; name::String = "")

MutexGroup contains one or more AbstractOptions and accepts command line arguments only if exactly one of the options was accepted.

source

Parsing Arguments

parse_args

CliOptions.parse_args — Function.

parse_args(spec::CliOptionSpec, args = ARGS)

Parse args according to the spec.

spec is an instance of CliOptionSpec which defines how to parse command line arguments. It is constructed with one or more concrete subtypes of AbstractOptions. See document of AbstractOption for full list of its subtypes.

args is the command line arguments to be parsed. If omitted, Base.ARGS – the command line arguments passed to the Julia script – will be parsed.

This function returns a ParseResult after parsing. It is basically a Dict-like object holding the values of options.

using CliOptions

spec = CliOptionSpec(
    Option(Int, "-n", "--num-workers"),
    FlagOption("-i", "--ignore-case"; negators = "--case-sensitive"),
    Positional("root"),
    Positional("pattern", "patterns"; multiple = true);
    program = "myfind"
)

args = parse_args(spec, split("-n 3 -i /var/log *.log", " "))
println("num_workers: ", args.num_workers)
println("ignore_case: ", args.ignore_case)
println("root: ", args.root)
println("patterns: ", args.patterns)

# output

num_workers: 3
ignore_case: true
root: /var/log
patterns: ["*.log"]

source

ParseResult

CliOptions.ParseResult — Type.

CliOptions.ParseResult()

Dict-like object holding parsing result of command line options. The values can be accessed using either:

dot notation (e.g.: result.num_workers)

bracket notation (e.g.: result["num_workers"])

This is the type parse_args function returns. If the function detected errors, it stores error messages into _errors field of this type. This may be useful if you let the program continue running on errors (see onerror parameter of CliOptionSpec).

source

update_defaults

CliOptions.update_defaults — Function.

update_defaults(result::ParseResult, defaults::Dict{String,Any})::ParseResult

Create a new ParseResult of which option values are updated by new default values.

ParseResult actually remembers the command line arguments and the default values which were originally defined by CliOptionSpec. This function firstly updates (merges) the default values stored in result using defaults, secondly resolves final option values, and finally creates and returns a new ParseResult.

This function is useful for a program which uses multiple sources of default values. For example, if you want to resolve option values in the following order:

Option values specified as command line argument

Option values read from a config file

Hard coded default value

you can use this function as below:

using CliOptions

# Firstly parse arguments normally
spec = CliOptionSpec(
    Option("--config-file"),
    Option("-x"; default = "foo"),
)
args = split("--config-file /path/to/config/file")
options = parse_args(spec, args)
println(options.x)  # We see hard-coded default value

# Let's pretend we loaded a config file and update defaults with it
config = Dict("x" => "bar")
options = update_defaults(options, config)
println(options.x)  # Now we see the default value in the config file

# If the option was specified in command line arguments, update_defaults has no effect
args = split("--config-file /path/to/config/file -x baz")
options = parse_args(spec, args)
options = update_defaults(options, config)
println(options.x)  # We see the value specified in the command line arguments

# output

foo
bar
baz

source

Error Handling

print_usage

CliOptions.print_usage — Function.

print_usage([io::IO], spec::CliOptionSpec; verbose = true)

Write usage (help) message to io. Set false to verbose if you want to print only the first line of the usage message. If io is omitted, message will be written stdout.

source

CliOptionError

CliOptions.CliOptionError — Type.

CliOptionError(msg::String)

An error occurred inside CliOptions module. Message describing the error is available in the msg field.

source

Internal Types

AbstractOption

CliOptions.AbstractOption — Type.

CliOptions.AbstractOption

Abstract supertype representing a command line option. Concrete subtypes are:

Option ... an option which takes a following argument as its value

FlagOption ... an option of which existence becomes its boolean value

CounterOption ... an option of which number of usage becomes its integer value

Positional ... an argument which is not an option

Note that a group of options represented with AbstractOptionGroup is also an AbstractOption so it can be used to construct CliOptionSpec.

source

AbstractOptionGroup

CliOptions.AbstractOptionGroup — Type.

CliOptions.AbstractOptionGroup

Abstract type representing a group of command line options. Concrete subtypes are:

OptionGroup

MutexGroup

source